edmdata: Data Sets for Psychometric Modeling

Collection of data sets from various assessments that can be used to evaluate psychometric models. These data sets have been analyzed in the following papers that introduced new methodology as part of the application section: Culpepper, S. A., & Balamuta, J. J. (2021) doi:10.1080/00273171.2021.1985949 , Yinghan Chen et al. (2021) doi:10.1007/s11336-021-09750-9 , Yinyin Chen et al. (2020) doi:10.1007/s11336-019-09693-2 , Culpepper, S. A. (2019a) doi:10.1007/s11336-019-09683-4 , Culpepper, S. A. (2019b) doi:10.1007/s11336-018-9643-8 , Culpepper, S. A., & Chen, Y. (2019) doi:10.3102/1076998618791306 , Culpepper, S. A., & Balamuta, J. J. (2017) doi:10.1007/s11336-015-9484-7 , and Culpepper, S. A. (2015) doi:10.3102/1076998615595403 .

See also

Author

Maintainer: James Joseph Balamuta balamut2@illinois.edu (ORCID) [copyright holder]

Authors:

Steven Andrew Culpepper sculpepp@illinois.edu (ORCID) [copyright holder]

Jeffrey Alan Douglas jeffdoug@illinois.edu [copyright holder]

Other contributors:

Auburn Jimenez aajimen2@illinois.edu [contributor, copyright holder]