Narcissistic Personality Inventory Item Responses

Narcissistic Personality Inventory Item Responses

items_narcissistic_personality_inventory

Format

An object of class matrix (inherits from array) with 11243 rows and 40 columns.

Details

Items with their desired option response bolded:

Q1

Option 1: I have a natural talent for influencing people

Option 2: I am not good at influencing people.

Q2

Option 1: Modesty doesn't become me

Option 2: I am essentially a modest person.

Q3

Option 1: I would do almost anything on a dare

Option 2: I tend to be a fairly cautious person.

Q4

Option 1: When people compliment me I sometimes get embarrassed

Option 2: I know that I am good because everybody keeps telling me so.

Q5

Option 1: The thought of ruling the world frightens the hell out of me

Option 2: If I ruled the world it would be a better place.

Q6

Option 1: I can usually talk my way out of anything

Option 2: I try to accept the consequences of my behavior.

Q7

Option 1: I prefer to blend in with the crowd

Option 2: I like to be the center of attention.

Q8

Option 1: I will be a success

Option 2: I am not too concerned about success.

Q9

Option 1: I am no better or worse than most people

Option 2: I think I am a special person.

Q10

Option 1: I am not sure if I would make a good leader

Option 2: I see myself as a good leader.

Q11

Option 1: I am assertive

Option 2: I wish I were more assertive.

Q12

Option 1: I like to have authority over other people

Option 2: I don't mind following orders.

Q13

Option 1: I find it easy to manipulate people

Option 2: I don't like it when I find myself manipulating people.

Q14

Option 1: I insist upon getting the respect that is due me

Option 2: I usually get the respect that I deserve.

Q15

Option 1: I don't particularly like to show off my body

Option 2: I like to show off my body.

Q16

Option 1: I can read people like a book

Option 2: People are sometimes hard to understand.

Q17

Option 1: If I feel competent I am willing to take responsibility for making decisions

Option 2: I like to take responsibility for making decisions.

Q18

Option 1: I just want to be reasonably happy

Option 2: I want to amount to something in the eyes of the world.

Q19

Option 1: My body is nothing special

Option 2: I like to look at my body.

Q20

Option 1: I try not to be a show off

Option 2: I will usually show off if I get the chance.

Q21

Option 1: I always know what I am doing

Option 2: Sometimes I am not sure of what I am doing.

Q22

Option 1: I sometimes depend on people to get things done

Option 2: I rarely depend on anyone else to get things done.

Q23

Option 1: Sometimes I tell good stories

Option 2: Everybody likes to hear my stories.

Q24

Option 1: I expect a great deal from other people

Option 2: I like to do things for other people.

Q25

Option 1: I will never be satisfied until I get all that I deserve

Option 2: I take my satisfactions as they come.

Q26

Option 1: Compliments embarrass me

Option 2: I like to be complimented.

Q27

Option 1: I have a strong will to power

Option 2: Power for its own sake doesn't interest me.

Q28

Option 1: I don't care about new fads and fashions

Option 2: I like to start new fads and fashions.

Q29

Option 1: I like to look at myself in the mirror

Option 2: I am not particularly interested in looking at myself in the mirror.

Q30

Option 1: I really like to be the center of attention

Option 2: It makes me uncomfortable to be the center of attention.

Q31

Option 1: I can live my life in any way I want to

Option 2: People can't always live their lives in terms of what they want.

Q32

Option 1: Being an authority doesn't mean that much to me

Option 2: People always seem to recognize my authority.

Q33

Option 1: I would prefer to be a leader

Option 2: It makes little difference to me whether I am a leader or not.

Q34

Option 1: I am going to be a great person

Option 2: I hope I am going to be successful.

Q35

Option 1: People sometimes believe what I tell them

Option 2: I can make anybody believe anything I want them to.

Q36

Option 1: I am a born leader

Option 2: Leadership is a quality that takes a long time to develop.

Q37

Option 1: I wish somebody would someday write my biography

Option 2: I don't like people to pry into my life for any reason.

Q38

Option 1: I get upset when people don't notice how I look when I go out in public

Option 2: I don't mind blending into the crowd when I go out in public.

Q39

Option 1: I am more capable than other people

Option 2: There is a lot that I can learn from other people.

Q40

Option 1: I am much like everybody else

Option 2: I am an extraordinary person.

Data pre-processing

We have applied list-wise deletion during pre-processing to remove any observations with missing values from the data set.

Answer coding

The subjects answered a set of assessment items seeking to determine the level of anxiety. Answers given in bold represent the desired response. If a subject matched this response, they were given a 1 inside of the item matrix, otherwise they received a zero.

References

Assessment Design:

Raskin, R., & Terry, H. (1988). A principal-components analysis of the narcissistic personality inventory and further evidence of its construct validity. Journal of Personality and Social Psychology, 54(5), 890. doi:10.1037/0022-3514.54.5.890

Data originated from:

OpenPsychometrics. 2013. Narcissistic Personality Inventory. https://openpsychometrics.org/_rawdata/NPI.zip.

Data used in:

TBA