Subset of Early Childhood Longitudinal Study, Kindergarten (ECLS-K)'s Approaches to Learning Item Responses

Subset of Early Childhood Longitudinal Study, Kindergarten (ECLS-K)'s Approaches to Learning Item Responses

items_ordered_eclsk_atl

Format

An object of class matrix (inherits from array) with 13354 rows and 12 columns.

Details

Items were split between being answered by Parents and Teachers.

Parents:

P4SRS10: Keep working at something until he/she is finished?

P4SRS13: Show interest in a variety of things?

P4SRS15: Concentrate on a task and ignore distractions?

P4SRS18: Help with chores?

P4SRS22: Eager to learn new things?

P4SRS24: Creative in work or in play?

Teachers:

T4SRS11: Keeps belongings organized.

T4SRS14: Shows eagerness to learn new things.

T4SRS15: Works independently.

T4SRS21: Easily adapts to changes in routine.

T4SRS23: Persists in completing tasks.

T4SRS24: Pays attention well.

Data pre-processing

The Early Childhood Longitudinal Study, Kindergarten (ECLS-K) has been subset down both the number of observations and variables. In particular, only observations without any missing values from a set of reduced variables -- given above -- are included. If additional data is required, please visit the data download page found in the reference section.

Answer coding

Parents and teachers each answered a set of survey items involving a likert scale that ranged from "1 = never" to "4 = very often" regarding the subject. Within the teacher responses, they also had the option of marking "-7 = no opportunity to observe" option, which was treated as a missing observation. To align with C++, we perform a index shift backward of 1 and, thus, make the scale "0=never" to "3=very often".

References

Data originated from:

NCES. (2010). Early childhood longitudinal study, kindergarten class of 1998-99 (ecls-k) kindergarten through fifth grade approaches to learning and self-description questionnaire (sdq) items and public-use data files. https://nces.ed.gov/pubsearch/pubsinfo.asp?pubid=2010070

Data used in:

Culpepper, S. (2019). An exploratory diagnostic model for ordinal responses with binary attributes: Identifiability and estimation. Psychometrika, 84(4), 921–940. doi: 10.1007/s11336-019-09683-4