Programme for International Student Assessment (PISA) 2012 U.S. Student Questionnaire Problem-Solving Vignettes

Programme for International Student Assessment (PISA) 2012 U.S. Student Questionnaire Problem-Solving Vignettes

items_ordered_pisa12_us_vignette

Format

An object of class matrix (inherits from array) with 3075 rows and 12 columns.

Details

Obtained from the PISA 2012 questionnaire responses given to 15-year olds, we selected 12 items under three vignettes and four item stems to measure problem-solving strategies applied by students.

Vignette (a):

Suppose that you have been sending text messages from your mobile phone for several weeks.Today, however, you can’t send text messages. You want to try to solve the problem.

ST96Q01: I press every button possible to find out what is wrong.

ST96Q02: I think about what might have caused the problem and what I can do to solve it.

ST96Q03: I read the manual.

ST96Q05: I ask a friend for help.

Vignette (b):

Suppose that you are planning a trip to the zoo with your brother. You don’t know which route to take to get there.

ST101Q01: I read the zoo brochure to see if it says how to get there.

ST101Q02: I study a map and work out the best route.

ST101Q03: I leave it to my brother to worry about how to get there.

ST101Q05: I know roughly where it is, so I suggest we just start driving.

Vignette (c):

Suppose that you arrive at the train station. There is a ticket machine that you have never used before. You want to buy a ticket.

ST104Q01: I check how similar it is to other ticket machines I have used.

ST104Q04: I try out all the buttons to see what happens.

ST104Q05: I ask someone for help.

ST104Q06: I try to find a ticket office at the station to buy a ticket.

Answer Coding

All items were reverse-coded such that responses categories begin with negative responses at 0 and go to 3 for a positive responses.

0: I would definitely not do this.

1: I would probably not do this.

2: I would probably do this.

3: I would definitely do this.

References

Assessment Design:

PISA Test Overview: https://www.oecd.org/pisa/aboutpisa/

PISA 2012 Mathematics Items: https://www.oecd.org/pisa/test/PISA%202012%20items%20for%20release_ENGLISH.pdf

Data used in:

Culpepper, S. A., & Balamuta, J. J. (2021). Inferring latent structure in polytomous data with a higher-order diagnostic model. Multivariate Behavioral Research, 1–19. doi:10.1080/00273171.2021.1985949