Trends in International Mathematics and Science Study 2015 (TIMSS) Grade 8 Student Background Survey Item Responses

Trends in International Mathematics and Science Study 2015 (TIMSS) Grade 8 Student Background Survey Item Responses

items_ordered_timss15_background

Format

An object of class matrix (inherits from array) with 9672 rows and 16 columns.

Details

From the school background data file, we selected 16 items under two distinct item stems that sought to measure students’ school experiences using a four-level likert-scale.

What do you think about your school? Tell me howmuch you agree with these statements:

"BSBG15A": I like being in school

"BSBG15B": I feel safe when I am in school

"BSBG15C": I feel like I belong at this school

"BSBG15D": I like to see my classmates at school.

"BSBG15E": Teachers at my school are fair to me.

"BSBG15F": I am proud to go to this school.

"BSBG15G": I learn a lot in school.

During this year, how often have the following things happened to you at school?

"BSBG16A": Made fun of me or called me names.

"BSBG16B": Left me out of their games or activities.

"BSBG16C": Spread lies about me.

"BSBG16D": Stole something from me.

"BSBG16E": Hit or hurt me (e.g. shoving, hitting, kicking).

"BSBG16F": Made me do things I didn’t want to.

"BSBG16G": Shared embarrassing information about me.

"BSBG16H": Posted embarrassing things about me online.

"BSBG16I": Threatened me.

Answer Coding

Items 8 through 16 were reverse-coded.

References

Data originated from:

Trends in International Mathematics and Science Study 2015 (TIMSS) Grade 8 Student Background Survey Item Responses (Mullis et al., 2016).

Data used in:

TBA