Programme for International Student Assessment (PISA) 2012 US Math Assessment

Programme for International Student Assessment (PISA) 2012 US Math Assessment

items_pisa12_us_math

Format

An object of class matrix (inherits from array) with 4978 rows and 76 columns.

Details

PISA assessment booklets have a spiral pattern. As a result, we've opted to re-order the data by booklet ID

Items alongside of the problem name:

pm00fq01: P2012 Apartment Purchase Q1

pm00gq01: P2012 An Advertising Column Q1

pm00kq02: P2012 Wheelchair Basketball Q2

pm033q01: P2000 A View with a Room Q1

pm034q01t: P2000 Bricks Q1

pm155q01: P2000 Pop Pyramids Q1

pm155q04t: P2000 Pop Pyramids Q4

pm192q01t: P2000 Containers Q1

pm273q01t: P2000 Pipelines Q1

pm305q01: P2000 Map Q1

pm406q01: P2003 Running Tracks Q1

pm406q02: P2003 Running Tracks Q2

pm408q01t: P2003 Lotteries Q1

pm411q01: P2003 Diving Q1

pm411q02: P2003 Diving Q2

pm420q01t: P2003 Transport Q1

pm423q01: P2003 Tossing Coins Q1

pm442q02: P2003 Braille Q2

pm446q01: P2003 The Thermometer Cricket Q1

pm446q02: P2003 The Thermometer Cricket Q2

pm447q01: P2003 Tile Arrangement Q1

pm464q01t: P2003 The Fence Q1

pm474q01: P2003 Running Time Q1

pm496q01t: P2003 Cash Withdrawal Q1

pm496q02: P2003 Cash Withdrawal Q2

pm559q01: P2003 Telephone Rates Q1

pm564q01: P2003 Chair Lift Q1

pm564q02: P2003 Chair Lift Q2

pm571q01: P2003 Stop the Car Q1

pm603q01t: P2003 Number Check Q1

pm800q01: P2003 Computer Game Q1

pm803q01t: P2003 Labels Q1

pm828q01: P2003 Carbon Dioxide Q1

pm828q02: P2003 Carbon Dioxide Q2

pm828q03: P2003 Carbon Dioxide Q3

pm903q03: P2012 Drip Rate Q3

pm905q01t: P2012 Tennis Balls Q1

pm905q02: P2012 Tennis Balls Q2

pm906q01: P2012 Crazy Ants Q1

pm909q01: P2012 Speeding Fines Q1

pm909q02: P2012 Speeding Fines Q2

pm909q03: P2012 Speeding Fines Q3

pm915q01: P2012 Carbon Tax Q1

pm915q02: P2012 Carbon Tax Q2

pm918q01: P2012 Charts Q1

pm918q02: P2012 Charts Q2

pm918q05: P2012 Charts Q5

pm919q01: P2012 Zs Fan Merchandise Q1

pm919q02: P2012 Zs Fan Merchandise Q2

pm923q01: P2012 Sailing Ships Q1

pm923q03: P2012 Sailing Ships Q3

pm923q04: P2012 Sailing Ships Q4

pm924q02: P2012 Sauce Q2

pm943q01: P2012 Arches Q1

pm943q02: P2012 Arches Q2

pm949q01t: P2012 Roof Truss Design Q1

pm949q02t: P2012 Roof Truss Design Q2

pm953q02: P2012 Flu Test Q2

pm953q03: P2012 Flu Test Q3

pm954q01: P2012 Medicine Doses Q1

pm954q02: P2012 Medicine Doses Q2

pm954q04: P2012 Medicine Doses Q4

pm955q01: P2012 Migration Q1

pm955q02: P2012 Migration Q2

pm982q01: P2012 Employment Data Q1

pm982q02: P2012 Employment Data Q2

pm982q03t: P2012 Employment Data Q3

pm982q04: P2012 Employment Data Q4

pm992q01: P2012 Spacers Q1

pm992q02: P2012 Spacers Q2

pm992q03: P2012 Spacers Q3

pm995q01: P2012 Revolving Door Q1

pm995q02: P2012 Revolving Door Q2

pm995q03: P2012 Revolving Door Q3

pm998q02: P2012 Bike Rental Q2

pm998q04t: P2012 Bike Rental Q4

Answer coding

The subjects answered a set of assessment items seeking to determine their knowledge of mathematics. The original PISA 2012 data set included variables that had problems with the following labels:

One level:

"N/A"

Four Levels:

"Score 0"

"Score 1"

"Score 7: N/A"

"Score 8: Not Attempted"

Five Levels:

"Score 0"

"Score 1"

"Score 2"

"Score 7: N/A"

"Score 8: Not Attempted"

Note, this data set includes items with four levels only. Further, items are coded with Score 0 as 0, Score 1 as 1, Score 7: N/A as NA, and Score 8: Not Attempted as 0.

References

Assessment Design:

PISA Test Overview: https://www.oecd.org/pisa/aboutpisa/

PISA 2012 Mathematics Items: https://www.oecd.org/pisa/test/PISA%202012%20items%20for%20release_ENGLISH.pdf

Data originated from:

Paul Bailey, Ahmad Emad, Huade Huo, Michael Lee, Yuqi Liao, Alex Lishinski, Trang Nguyen, Qingshu Xie, Jiao Yu, Ting Zhang, Eric Buehler, Jeppe Bundsgaard, Ren C'deBaca and Anders Astrup Christensen (2021). EdSurvey: Analysis of NCES Education Survey and Assessment Data. R package version 2.7.0. https://CRAN.R-project.org/package=EdSurvey

Data used in:

TBA