Last Series of the Standard Progressive Matrices (SPM-LS) Item Responses

Last Series of the Standard Progressive Matrices (SPM-LS) Item Responses

items_spm_ls

Format

An object of class matrix (inherits from array) with 499 rows and 12 columns.

Details

Items with the correct answer response based off of Table 9 of the Robitzsch (2020) pre-print paper.

SPM1: 7

SPM2: 6

SPM3: 8

SPM4: 2

SPM5: 1

SPM6: 5

SPM7: 1

SPM8: 6

SPM9: 3

SPM10: 2

SPM11: 4

SPM12: 5

Answer coding

The subjects answered a set of assessment items seeking to determine the level of matrix reasoning. Answers given in bold represent the desired response. If a subject matched this response, they were given a 1 inside of the item matrix, otherwise they received a zero.

References

Assessment Design:

Raven, J. C. (1941). Standardization of progressive matrices, 1938. British Journal of Medical Psychology, 19(1), 137–150. doi: 10.1111/j.2044-8341.1941.tb00316.x

Data originated from:

Myszkowski, N., & Storme, M. (2018). A snapshot of g? Binary and polytomous item-response theory investigations of the last series of the standard progressive matrices (spm-ls). Intelligence, 68, 109–116. doi: 10.1016/j.intell.2018.03.010

Robitzsch, A. (2020). Regularized latent class analysis for polytomous item responses: An application to spm-ls data. Preprint. doi: 10.20944/preprints202007.0269.v1

Data used in:

TBA