Taylor Manifest Anxiety Scale Item Responses

Taylor Manifest Anxiety Scale Item Responses

items_taylor_manifest_anxiety_scale

Format

An object of class matrix (inherits from array) with 4468 rows and 50 columns.

Details

Questions alongside of their correct answer is based off of Table 1 of the Taylor (1953) paper.

Items with their desired response bolded:

Q1: I do not tire quickly. (False)

Q2: I am troubled by attacks of nausea. (True)

Q3: I believe I am no more nervous than most others. (False)

Q4: I have very few headaches. (False)

Q5: I work under a great deal of tension. (True)

Q6: I cannot keep my mind on one thing. (True)

Q7: I worry over money and business. (True)

Q8: I frequently notice my hand shakes when I try to do something. (True)

Q9: I blush no more often than others. (False)

Q10: I have diarrhea once a month or more. (True)

Q11: I worry quite a bit over possible misfortunes. (True)

Q12: I practically never blush. (False)

Q13: I am often afraid that I am going to blush. (True)

Q14: I have nightmares every few nights. (True)

Q15: My hands and feet are usually warm. (False)

Q16: I sweat very easily even on cool days. (True)

Q17: Sometimes when embarrassed, I break out in a sweat. (True)

Q18: I hardly ever notice my heart pounding and I am seldom short of breath. (False)

Q19: I feel hungry almost all the time. (True)

Q20: I am very seldom troubled by constipation. (False)

Q21: I have a great deal of stomach trouble. (True)

Q22: I have had periods in which I lost sleep over worry. (True)

Q23: My sleep is fitful and disturbed. (True)

Q24: I dream frequently about things that are best kept to myself. (True)

Q25: I am easily embarrassed. (True)

Q26: I am more sensitive than most other people. (True)

Q27: I frequently find myself worrying about something. (True)

Q28: I wish I could be as happy as others seem to be. (True)

Q29: I am usually calm and not easily upset. (False)

Q30: I cry easily. (True)

Q31: I feel anxiety about something or someone almost all the time. (True)

Q32: I am happy most of the time. (False)

Q33: It makes me nervous to have to wait. (True)

Q34: I have periods of such great restlessness that I cannot sit long I a chair. (True)

Q35: Sometimes I become so excited that I find it hard to get to sleep. (True)

Q36: I have sometimes felt that difficulties were piling up so high that I could not overcome them. (True)

Q37: I must admit that I have at times been worried beyond reason over something that really did not matter. (True)

Q38: I have very few fears compared to my friends. (False)

Q39: I have been afraid of things or people that I know could not hurt me. (True)

Q40: I certainly feel useless at times. (True)

Q41: I find it hard to keep my mind on a task or job. (True)

Q42: I am usually self-conscious. (True)

Q43: I am inclined to take things hard. (True)

Q44: I am a high-strung person. (True)

Q45: Life is a trial for me much of the time. (True)

Q46: At times I think I am no good at all. (True)

Q47: I am certainly lacking in self-confidence. (True)

Q48: I sometimes feel that I am about to go to pieces. (True)

Q49: I shrink from facing crisis of difficulty. (True)

Q50: I am entirely self-confident. (False)

Data pre-processing

We have applied list-wise deletion during pre-processing to remove any observations with missing values from the data set.

Answer coding

The subjects answered a set of assessment items seeking to determine the level of anxiety. Answers given in bold represent the desired response. If a subject matched this response, they were given a 1 inside of the item matrix, otherwise they received a zero.

References

Assessment Design:

Taylor, J. A. (1953). A personality scale of manifest anxiety. The Journal of Abnormal and Social Psychology, 48(2), 285–290. doi:10.1037/h0056264 .

Data originated from:

OpenPsychometrics. 2012. Taylor Manifest Anxiety Scale. https://openpsychometrics.org/_rawdata/TMA.zip.

Data used in:

TBA