Fraction Subtraction and Addition Assessment Expert-Derived Q Matrix

Fraction Subtraction and Addition Assessment Expert-Derived Q Matrix

qmatrix_fractions

Format

An object of class matrix (inherits from array) with 20 rows and 8 columns.

Details

Each entry in the matrix is either 1, if the Item uses the Trait, or 0, if the Item does not use the Trait. The traits identified by this matrix are:

Trait1: Convert a whole number to a fraction,

Trait2: Separate a whole number from fraction,

Trait3: Simplify before subtraction,

Trait4: Find a common denominator,

Trait5: Borrow from the whole number part,

Trait6: Column borrow to subtract the second numerator from the first,

Trait7: Subtract numerators,

Trait8: Reduce answers to simplest form.

The subjects answered the following assessment items:

Item01: \(\frac{5}{3}-\frac{3}{4}\)

Item02: \(\frac{3}{4}-\frac{3}{8}\)

Item03: \(\frac{5}{6}-\frac{1}{9}\)

Item04: \(3\frac{1}{2}-2\frac{3}{2}\)

Item05: \(4\frac{3}{5}-3\frac{4}{10}\)

Item06: \(\frac{6}{7}-\frac{4}{7}\)

Item07: \(3-2\frac{1}{5}\)

Item08: \(\frac{2}{3}-\frac{2}{3}\)

Item09: \(3\frac{7}{8}-2\)

Item10: \(4\frac{4}{12}-2\frac{7}{12}\)

Item11: \(4\frac{1}{3}-2\frac{4}{3}\)

Item12: \(\frac{11}{8}-\frac{1}{8}\)

Item13: \(3\frac{3}{8}-2\frac{5}{6}\)

Item14: \(3\frac{4}{5}-3\frac{2}{5}\)

Item15: \(2-\frac{1}{3}\)

Item16: \(4\frac{5}{7}-1\frac{4}{7}\)

Item17: \(7\frac{3}{5}-2\frac{4}{5}\)

Item18: \(4\frac{1}{10}-2\frac{8}{10}\)

Item19: \(4-1\frac{4}{3}\)

Item20: \(4\frac{1}{3}-1\frac{5}{3}\)

References

Data originated from:

Tatsuoka, C. (2002). Data analytic methods for latent partially ordered classification models. Journal of the Royal Statistical Society: Series C (Applied Statistics), 51(3), 337–350. doi: 10.1111/1467-9876.00272

Tatsuoka, K. K. (1984), Analysis of errors in fraction addition and subtraction problems (Final Report for Grant No. NIE-G-81-0002). Urbana: University of Illinois, Computer-Based Education Research Laboratory (CERL).

Data used in:

Chen, Y., Liu, Y., Culpepper, S. A., & Chen, Y. (2021). Inferring the number of attributes for the exploratory DINA model. Psychometrika, 86(1), 30–64. doi: 10.1007/s11336-021-09750-9

Chen, Y., Culpepper, S. A., & Liang, F. (2020). A sparse latent class model for cognitive diagnosis. Psychometrika, 1–33. doi: 10.1007/s11336-019-09693-2

Culpepper, S. A. (2019). Estimating the cognitive diagnosis \(Q\) matrix with expert knowledge: Application to the fraction-subtraction dataset. Psychometrika, 84(2), 333–357. doi: 10.1007/s11336-018-9643-8

Culpepper, S. A., & Chen, Y. (2019). Development and application of an exploratory reduced reparameterized unified model. Journal of Educational and Behavioral Statistics, 44(1), 3–24. doi: 10.3102/1076998618791306

Chen, Y., Culpepper, S. A., Chen, Y., & Douglas, J. (2018). Bayesian estimation of the dina q matrix. Psychometrika, 83(1), 89–108. doi: 10.1007/s11336-017-9579-4