Elementary Probability Theory Assessment Expert-Derived Q Matrix

Elementary Probability Theory Assessment Expert-Derived Q Matrix

qmatrix_probability_part_one

Format

An object of class matrix (inherits from array) with 12 rows and 4 columns.

Details

Each entry in the matrix is either 1, if the item uses the trait, or 0, if the item does not use the trait. The traits identified by this matrix are:

cp: the probability of the complement of an event

id: two independent events

pb: probability of an event

un: union of two disjoint events

For a detailed overview of items, please see items_probability_part_one_full or items_probability_part_one_reduced.

Identifiability

Note, the expert supplied Q-matrix is not strictly identified. Though, the expert matrix is generically identified.

References

Data originated from:

Heller, J., & Wickelmaier, F. (2013). Minimum discrepancy estimation in probabilistic knowledge structures. Electronic Notes in Discrete Mathematics, 42, 49–56. doi:10.1016/j.endm.2013.05.145

Data used in:

Chen, Yinghan, Liu, Y., Culpepper, S. A., & Chen, Y. (2021). Inferring the number of attributes for the exploratory DINA model. Psychometrika, 86(1), 30–64. doi:10.1007/s11336-021-09750-9